siteId split
