+++ src/packages/universalLoginPage/constants.ts
@@ -1,5 +1,3 @@
-import type { Site, Sites } from "./types";
-
 const SITE_LOGOS_VERSION_KEY = "_v_";
 export const LOGO_IMG_ID = "prompt-logo-center";
 
@@ -12,62 +10,4 @@ export const CONNECTOR_LOGO_BASE_URL =
   "https://raw.githubusercontent.com/sesam-io";
 export const BASE_LOGO_URL = `https://cdn.jsdelivr.net/gh/sesam-io/wix-sso@${SITE_LOGOS_RELEASE_VERSION}/src/packages/siteLogos`;
 
-const DEFAULT_LOGO_URL = `${BASE_LOGO_URL}/sesam/sesam-talk-rgb.png`;
-
-const tripletexSite: Site = {
-  id: "tripletex",
-  logoUrl: `${CONNECTOR_LOGO_BASE_URL}/tripletex-connector/main/assets/tripletex-logo-horizontal-complete.svg`,
-  loginSubTitle: "Log in to Tripletex DataSync.",
-  signupSubTitle: "Sign up to Tripletex DataSync.",
-};
-
-const superofficeSite: Site = {
-  id: "superoffice",
-  logoUrl: `${BASE_LOGO_URL}/superoffice/superoffice-ulp-header-logo.svg`,
-  loginSubTitle: "Log in to SuperOffice DataSync.",
-  signupSubTitle: "Sign up to SuperOffice DataSync.",
-  titleClassName: "superofficeLogInTitle",
-};
-
-export const WixSites: Sites = {
-  hubspot: {
-    id: "hubspot",
-    logoUrl: `${BASE_LOGO_URL}/hubspot/making-hubspot-talk-logo-centered.svg`,
-    loginSubTitle: "Log in to Making HubSpot Talk.",
-    signupSubTitle: "Sign up to Making HubSpot Talk.",
-  },
-  powerofficego: {
-    id: "powerofficego",
-    logoUrl: `${BASE_LOGO_URL}/poweroffice/poweroffice-new-logo.png`,
-    loginSubTitle: "Log in to PowerOffice DataSync.",
-    signupSubTitle: "Sign up to PowerOffice DataSync.",
-  },
-  sesam: {
-    id: "sesam",
-    logoUrl: DEFAULT_LOGO_URL,
-    loginSubTitle: "Log in to Sesam Talk.",
-    signupSubTitle: "Sign up to Sesam Talk to continue to Sesam Talk.",
-  },
-  superoffice: {
-    id: "superoffice",
-    logoUrl: `${BASE_LOGO_URL}/superoffice/superoffice-ulp-header-logo.svg`,
-    loginSubTitle: "Log in to SuperOffice DataSync.",
-    signupSubTitle: "Sign up to SuperOffice DataSync.",
-    titleClassName: "superofficeLogInTitle",
-  },
-  "superoffice-test": {
-    ...superofficeSite,
-    id: "superoffice-test",
-  },
-  tripletex: tripletexSite,
-  "tripletex-test": {
-    ...tripletexSite,
-    id: "tripletex-test",
-  },
-  wave: {
-    id: "wave",
-    logoUrl: `${BASE_LOGO_URL}/wave/making-wave-talk-logo-centered.svg`,
-    loginSubTitle: "Log in to Making Wave Talk.",
-    signupSubTitle: "Sign up to Making Wave Talk.",
-  },
-};
+export const DEFAULT_LOGO_URL = `${BASE_LOGO_URL}/sesam/sesam-talk-rgb.png`;
